♻️ Refactor server architecture with new handlers
- Add dedicated API and TDD handlers
- Replace old server/index.js with new http-server.js
- Improve separation of concerns
